services: Move operation mapping to operationToTransactionMapper
